refactor out a ButtonManager from the CommandManager for better single responsibility principle
